start, reorganize balancerd
* replace main.rs with balancerd.rs, flatten service.rs
