+++ src/balancerd/src/lib.rs
@@ -18,6 +18,7 @@
 
 mod codec;
 mod dyncfgs;
+// mod service;
 
 use std::collections::BTreeMap;
 use std::net::SocketAddr;
